Adjusted the TP command because it interferes with the built-in one, wich is more powerful now.
Created a sethome command, allowing a person to set their home point. Unfortunately it needs some work, because it doesn't seem to work after death because there is no bed at the location.

Added a command to grant 5 command blocks, beacuse there is no in-game way to get them if you don't know the ID number.